Tackle issues of integer width and sign.
* Enable compiler warnings for potentially problematic implicit casts
* Go through code and add explicit casts or avoid conversions
* Switch to using 64-bit integers for adlb IDs and integers on all platforms (i.e. use long long instead of long)
* Also switch data structures to long long to match.
* Add in typedefs for adlb IDs and integers, use the typedefed versions instead of the plain long type
* Use WideInt in tcl/c interface code

git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/lb/trunk@7269 dc4e9af1-7f46-4ead-bba6-71afc04862de

+++ code/src/messaging.c
@@ -35,7 +35,7 @@
 static char* tag_names[XLB_MAX_TAGS];
 
 static char* tag_prefix;
-static int tag_prefix_length;
+static size_t tag_prefix_length;
 
 static void add_tags(void);
